Rename "install" to "finalize"
Make it more clear that this runs at the end
